Click and Rotate, Scale. Fix undo, redo. Send Back, Front, Forward, Backward

+++ ShapeLibrary/ITransformer.cs
@@ -6,7 +6,7 @@ namespace ShapeLibrary
     {
         void Draw(Graphics graphics);
         void Fill(Graphics graphics);
-        void Rotate(Graphics graphics, int angle);
+        void Rotate(Graphics graphics, float angle);
         void Scale(Graphics graphics, float xRate, float yRate);
         void Shift(Graphics graphics, int dx, int dy);
     }
